building walls

+++ raycasting/draw_projection_of_view.c
@@ -6,7 +6,7 @@
 /*   By: iassil <iassil@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/05/07 21:42:43 by iassil            #+#    #+#             */
-/*   Updated: 2024/05/24 15:32:39 by iassil           ###   ########.fr       */
+/*   Updated: 2024/05/24 21:48:17 by iassil           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -55,8 +55,8 @@ void	ft_draw_line_of_view(t_cube *data, int color)
 {
 	t_line		l;
 
-	l.x_begin = (data->player.x + ((double)PLAYER_PX / 2)) * MINIMAP;
-	l.y_begin = (data->player.y + ((double)PLAYER_PX / 2)) * MINIMAP;
+	l.x_begin = (data->player.x + ((float)PLAYER_PX / 2)) * MINIMAP;
+	l.y_begin = (data->player.y + ((float)PLAYER_PX / 2)) * MINIMAP;
 	l.x_end = (l.x_begin + cos(data->player.rotation_angle) * LINE) * MINIMAP;
 	l.y_end = (l.y_begin + sin(data->player.rotation_angle) * LINE) * MINIMAP;
 	ft_draw_line(data, l, color);
@@ -76,8 +76,8 @@ void	ft_draw_line_of_view_60deg(t_cube *data, int color)
 		ra -= 2 * M_PI;
 	while (i <= NUM_RAYS)
 	{
-		l.x_begin = (data->player.x + ((double)PLAYER_PX / 2)) * MINIMAP;
-		l.y_begin = (data->player.y + ((double)PLAYER_PX / 2)) * MINIMAP;
+		l.x_begin = (data->player.x + ((float)PLAYER_PX / 2)) * MINIMAP;
+		l.y_begin = (data->player.y + ((float)PLAYER_PX / 2)) * MINIMAP;
 		l.x_end = (l.x_begin + cos(ra) * LINE) * MINIMAP;
 		l.y_end = (l.y_begin + sin(ra) * LINE) * MINIMAP;
 		ft_draw_line(data, l, color);
